+++ src/test/scala/QuartzSchedulerFunctionalSpec.scala
@@ -1,28 +1,26 @@
 package com.typesafe.akka.extension.quartz
 
-import java.util.{Calendar, Date}
-
-import akka.japi.Option.Some
-import org.junit.runner.RunWith
-import com.typesafe.config.ConfigFactory
 import akka.actor._
+import akka.japi.Option.Some
 import akka.testkit._
+import org.junit.runner.RunWith
+import org.scalatest.BeforeAndAfterAll
+import org.scalatest.matchers.must.Matchers
+import org.scalatest.wordspec.AnyWordSpecLike
+import org.specs2.runner.JUnitRunner
+
+import java.util.{Calendar, Date}
 import scala.concurrent._
 import scala.concurrent.duration._
-import org.scalatest.junit.JUnitRunner
-import org.scalatest.{BeforeAndAfterAll, WordSpecLike}
-import org.scalatest.MustMatchers
-import org.scalatest.Matchers
-
 
 @RunWith(classOf[JUnitRunner])
 class QuartzSchedulerFunctionalSpec(_system: ActorSystem) extends TestKit(_system: ActorSystem)
   with ImplicitSender
-  with WordSpecLike
-  with MustMatchers
+  with AnyWordSpecLike
+  with Matchers
   with BeforeAndAfterAll {
 
-  override def afterAll {
+  override protected def afterAll(): Unit = {
     system.terminate()
     Await.result(system.whenTerminated, Duration.Inf)
   }
@@ -172,9 +170,11 @@ class QuartzSchedulerFunctionalSpec(_system: ActorSystem) extends TestKit(_syste
       receipt must contain(Tock)
       receipt must have size (5)
     }
+
   }
 
   "The Quartz Scheduling Extension with Reschedule" must {
+
     "Reschedule an existing Cron Job" in {
       val receiver = _system.actorOf(Props(new ScheduleTestReceiver))
       val probe = TestProbe()
@@ -188,6 +188,7 @@ class QuartzSchedulerFunctionalSpec(_system: ActorSystem) extends TestKit(_syste
         jobCalender.get(Calendar.SECOND) mustEqual 59
       }
     }
+
   }
 
   "Get next trigger date by schedule name" in {
@@ -201,6 +202,7 @@ class QuartzSchedulerFunctionalSpec(_system: ActorSystem) extends TestKit(_syste
   }
 
   "The Quartz Scheduling Extension with Dynamic Create" must {
+
     "Throw exception if creating schedule that already exists" in {
       val receiver = _system.actorOf(Props(new ScheduleTestReceiver))
 
@@ -235,8 +237,8 @@ class QuartzSchedulerFunctionalSpec(_system: ActorSystem) extends TestKit(_syste
       receipt must contain(Tock)
       receipt must have size (5)
     }
-  }
 
+  }
 
   /**
    * JobSchedule operations {create, update, delete} combine existing
@@ -245,6 +247,7 @@ class QuartzSchedulerFunctionalSpec(_system: ActorSystem) extends TestKit(_syste
    * consistency with existing rescheduleJob method).
    */
   "The Quartz Scheduling Extension with Dynamic create, update, delete JobSchedule operations" must {
+
     "Throw exception if creating job schedule that already exists" in {
 
       val alreadyExistingScheduleJobName = "cronEvery10Seconds"
@@ -306,7 +309,6 @@ class QuartzSchedulerFunctionalSpec(_system: ActorSystem) extends TestKit(_syste
       }
     }
 
-
     "Delete an existing job schedule Cron Job without any error and allow successful creation of new schedule with identical job name" in {
 
       val toDeleteSheduleJobName = "toBeDeletedscheduleCron_1"
@@ -351,10 +353,7 @@ class QuartzSchedulerFunctionalSpec(_system: ActorSystem) extends TestKit(_syste
       }
     }
 
-
-
   }
-
 
   case class NewProbe(probe: ActorRef)
   case object Tick
@@ -378,99 +377,4 @@ class QuartzSchedulerFunctionalSpec(_system: ActorSystem) extends TestKit(_syste
     }
   }
 
-}
-
-object SchedulingFunctionalTest {
-    lazy val sampleConfiguration = { ConfigFactory.parseString("""
-    akka {
-      event-handlers = ["akka.testkit.TestEventListener"]
-      loglevel = "INFO"
-      quartz {
-        defaultTimezone = "UTC"
-        schedules {
-          cronEvery30Seconds {
-            description = "A cron job that fires off every 30 seconds"
-            expression = "*/30 * * ? * *"
-          }
-          cronEvery15Seconds {
-            description = "A cron job that fires off every 15 seconds"
-            expression = "*/15 * * ? * *"
-          }
-          cronEvery12Seconds {
-            description = "A cron job that fires off every 10 seconds"
-            expression = "*/12 * * ? * *"
-          }
-          cronEvery10SecondsWithFireTime{
-          description = "A cron job that fires off every 10 seconds with FireTime"
-          expression = "*/10 * * ? * *"}
-          cronEvery10Seconds {
-            description = "A cron job that fires off every 10 seconds"
-            expression = "*/10 * * ? * *"
-          }
-          cronEvery10SecondsWithFireTimes{
-          description = "A cron job that fires off every 10 seconds with FireTimes"
-          expression = "*/10 * * ? * *"}
-          cronEvery10Seconds {
-            description = "A cron job that fires off every 10 seconds"
-            expression = "*/10 * * ? * *"
-          }
-          cronEvery5Seconds {
-            description = "A cron job that fires off every 5 seconds"
-            expression = "*/5 * * ? * *"
-          }
-          cronEveryEvenSecond {
-            description = "A cron job that fires off every even second"
-            expression = "0/2 * * ? * *"
-          }
-          cronEveryMidnight {
-            description = "A cron job that fires off every Midnight"
-            expression = "0 0 0 * * ?"
-          }
-        }
-        calendars {
-          WinterClosings {
-            type = Annual
-            description = "Major holiday dates that occur in the winter time every year, non-moveable (The year doesn't matter)"
-            excludeDates = ["12-25", "01-01"]
-          }
-          Easter {
-            type = Holiday
-            description = "The easter holiday (a moveable feast) for the next five years"
-            excludeDates = ["2013-03-31", "2014-04-20", "2015-04-05", "2016-03-27", "2017-04-16"]
-          }
-          HourOfTheWolf {
-            type = Daily
-            description = "A period every day in which cron jobs are quiesced, during night hours"
-            exclude {
-              startTime = "03:00"
-              endTime   = "05:00:00"
-            }
-            timezone = PST
-          }
-          FirstOfMonth {
-            type = Monthly
-            description = "A thinly veiled example to test monthly exclusions of one day"
-            excludeDays = [1]
-          }
-          FirstAndLastOfMonth {
-            type = Monthly
-            description = "A thinly veiled example to test monthly exclusions"
-            excludeDays = [1, 31]
-          }
-          MondaysSuck {
-            type = Weekly
-            description = "Everyone, including this calendar, hates mondays as an integer"
-            excludeDays = [2]
-            excludeWeekends = false
-          }
-          CronOnlyBusinessHours {
-            type = Cron
-            excludeExpression = "* * 0-7,18-23 ? * *"
-            timezone = "America/Los_Angeles"
-          }
-        }
-      }
-    }
-                                                               """.stripMargin)
-  }
 }